Kill some more compiler warnings under Linux and mingw32
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@5031 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/InstallOptions/InstallerOptions.cpp
@@ -1090,7 +1090,7 @@ int WINAPI createCfgDlg()
 
 #undef DEFAULT_STYLES
 
-    if (pField->nType < 1 || pField->nType > (sizeof(ClassTable) / sizeof(ClassTable[0])))
+    if (pField->nType < 1 || pField->nType > (int)(sizeof(ClassTable) / sizeof(ClassTable[0])))
       continue;
 
     DWORD dwStyle, dwExStyle;
